cmd/bent: exclude bss in size total
Go commit fc88e18 added a substantial .noptrbss region to every program
that includes the crypto packages. `size` reports this region as part of
the total size of the binary, even though this section isn't realized in
the ELF binary itself and is demand-paged only when FIPS 140-3 mode is
enabled.

Since we're interested in how big the binaries are on-disk and
in-memory, grab the total size via stat. This should make this benchmark
useful again.

Change-Id: I05982932f2b2a6b9b31e737a0a8d90f0029564ef
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/727980
Reviewed-by: David Chase <drchase@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>